Fix export_to_tsv exporting only the first parameter for dynamic relations
Some operators (affine, complex_diagonal) have more than one parameters,
however export_to_tsv was only exporting the first one. This caused, for
example, the ComplEx FB15k export to have only 200 params per relation
(the real part), with the other 200 (the imaginary part) being dropped.

Tested using the FB15k and (a modified version of) the LiveJournal
example scripts.

+++ torchbiggraph/converters/utils.py
@@ -22,6 +22,12 @@ def extract_gzip(gzip_path: str, remove_finished: bool = False) -> str:
     if ext != ".gz":
         raise RuntimeError("Not a gzipped file")
 
+    if os.path.exists(fpath):
+        print("Found a file that indicates that the input data "
+              "has already been extracted, not doing it again.")
+        print("This file is: %s" % fpath)
+        return fpath
+
     with open(fpath, "wb") as out_bf, gzip.GzipFile(gzip_path) as zip_f:
         shutil.copyfileobj(zip_f, out_bf)
     if remove_finished:
